+++ src/components/Header.js
@@ -1,0 +1,44 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import md5 from 'crypto-js/md5';
+import { connect } from 'react-redux';
+
+class Header extends React.Component {
+  render() {
+    const { count, name, email } = this.props;
+    const hashEmail = md5(email).toString();
+    return (
+      <header>
+        <img
+          src={ `https://www.gravatar.com/avatar/${hashEmail}` }
+          alt="foto"
+          data-testid="header-profile-picture"
+        />
+        <span
+          data-testid="header-player-name"
+        >
+          { name }
+        </span>
+        <span
+          data-testid="header-score"
+        >
+          { count }
+        </span>
+      </header>
+    );
+  }
+}
+
+const mapStateToProps = (state) => ({
+  count: state.player.score,
+  name: state.player.name,
+  email: state.player.gravatarEmail,
+});
+
+export default connect(mapStateToProps)(Header);
+
+Header.propTypes = {
+  count: PropTypes.number.isRequired,
+  name: PropTypes.string.isRequired,
+  email: PropTypes.string.isRequired,
+};
